CMDCT-4189: moving form schema to a new file
