+++ services/ui-src/src/components/fields/TextField.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { FieldError, useFormContext } from "react-hook-form";
+import { get, useFormContext } from "react-hook-form";
 import { TextField as CmsdsTextField } from "@cmsgov/design-system";
 import { Box } from "@chakra-ui/react";
 import { parseCustomHtml } from "utils";
@@ -32,11 +32,8 @@ export const TextField = (props: PageElementProps) => {
   };
 
   // prepare error message, hint, and classes
-  const formErrorState = form?.formState?.errors;
-  const elementErrors = formErrorState?.[props.formkey] as {
-    answer: FieldError;
-  };
-  const errorMessage = elementErrors?.answer?.message;
+  const formErrors = form?.formState?.errors;
+  const errorMessage: string | undefined = get(formErrors, key)?.message;
   const parsedHint = textbox.helperText && parseCustomHtml(textbox.helperText);
   const labelText = textbox.label;
